fix issue and rename manifest
