+++ lib/cli-engine/formatters/_manifest.js
